Add front end justifcation Upgrade mpost Add other dynamic axes in addition to leftTatweel and rightTatweel

+++ lib/mplib/mpsvgout.h
@@ -1,5 +1,5 @@
 /*3:*/
-// #line 80 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 78 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 #ifndef MPSVGOUT_H
 #define MPSVGOUT_H 1
@@ -8,51 +8,53 @@
 #include "mplibps.h"
 typedef struct svgout_data_struct{
 /*6:*/
-// #line 113 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 111 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 size_t file_offset;
 
 /*:6*//*12:*/
-// #line 165 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 163 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 char*buf;
 unsigned loc;
 unsigned bufsize;
 
 /*:12*//*20:*/
-// #line 299 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 297 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 int level;
 
 /*:20*//*27:*/
-// #line 386 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 384 "../../../source/texk/web2c/mplibdir/svgout.w"
 
-integer dx;
-integer dy;
+double dx;
+double dy;
+
 
 /*:27*//*56:*/
-// #line 1159 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 1158 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 int clipid;
 
 /*:56*/
-// #line 87 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 85 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 }svgout_data_struct;
 /*4:*/
-// #line 92 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 90 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 void mp_svg_backend_initialize(MP mp);
 void mp_svg_backend_free(MP mp);
 
 /*:4*//*60:*/
-// #line 1209 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 1208 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 int mp_svg_gr_ship_out(mp_edge_object*hh,int prologues,int standalone);
 
 /*:60*/
-// #line 89 "../../../source/texk/web2c/mplibdir/svgout.w"
+#line 87 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 #endif
+#line 89 "../../../source/texk/web2c/mplibdir/svgout.w"
 
 /*:3*/
